+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/ImplementationBridgeHelpers.java
@@ -41,7 +41,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setCosmosClientBuilderAccessor(final CosmosClientBuilderAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosClientBuilderHelper accessor already initialized!");
+                throw new IllegalStateException("CosmosClientBuilder accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -49,7 +49,7 @@ public class ImplementationBridgeHelpers {
 
         static CosmosClientBuilderAccessor getCosmosClientBuilderAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosClientBuilderHelper accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosClientBuilder accessor is not initialized yet!");
             }
 
             return accessor;
@@ -73,7 +73,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setPartitionKeyAccessor(final PartitionKeyAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("PartitionKeyHelper accessor already initialized!");
+                throw new IllegalStateException("PartitionKey accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -81,7 +81,7 @@ public class ImplementationBridgeHelpers {
 
         public static PartitionKeyAccessor getPartitionKeyAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("PartitionKeyHelper accessor is not initialized!");
+                throw new IllegalStateException("PartitionKey accessor is not initialized!");
             }
 
             return accessor;
@@ -102,7 +102,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setCosmosQueryRequestOptionsAccessor(final CosmosQueryRequestOptionsAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor already initialized!");
+                throw new IllegalStateException("CosmosQueryRequestOptions accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -110,7 +110,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosQueryRequestOptionsAccessor getCosmosQueryRequestOptionsAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosQueryRequestOptions accessor is not initialized yet!");
             }
 
             return accessor;
@@ -162,12 +162,12 @@ public class ImplementationBridgeHelpers {
 
         private CosmosBulkProcessingOptionsHelper() {}
         static {
-            ensureClassLoaded(CosmosQueryRequestOptionsHelper.class);
+            ensureClassLoaded(BulkProcessingOptions.class);
         }
 
         public static void setCosmosBulkProcessingOptionAccessor(final CosmosBulkProcessingOptionAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor already initialized!");
+                throw new IllegalStateException("BulkProcessingOptions accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -175,7 +175,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosBulkProcessingOptionAccessor getCosmosBulkProcessingOptionAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosQueryRequestOptionsHelper accessor is not initialized yet!");
+                throw new IllegalStateException("BulkProcessingOptions accessor is not initialized yet!");
             }
 
             return accessor;
@@ -199,7 +199,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setCosmosItemResponseBuilderAccessor(final CosmosItemResponseBuilderAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosItemResponseBuilder accessor already initialized!");
+                throw new IllegalStateException("CosmosItemResponse accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -207,7 +207,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosItemResponseBuilderAccessor getCosmosItemResponseBuilderAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosItemResponseBuilder accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosItemResponse accessor is not initialized yet!");
             }
 
             return accessor;
@@ -238,7 +238,7 @@ public class ImplementationBridgeHelpers {
 
         public static void setCosmosClientAccessor(final CosmosClientAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("CosmosClientAccessor accessor already initialized!");
+                throw new IllegalStateException("CosmosClient accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -246,7 +246,7 @@ public class ImplementationBridgeHelpers {
 
         public static CosmosClientAccessor geCosmosClientAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("CosmosClientAccessor accessor is not initialized yet!");
+                throw new IllegalStateException("CosmosClient accessor is not initialized yet!");
             }
 
             return accessor;
@@ -358,12 +358,12 @@ public class ImplementationBridgeHelpers {
         }
 
         static {
-            ensureClassLoaded(BulkProcessingThresholdsHelper.class);
+            ensureClassLoaded(BulkProcessingThresholds.class);
         }
 
         public static void setBulkProcessingThresholdsAccessor(final BulkProcessingThresholdsAccessor newAccessor) {
             if (accessor != null) {
-                throw new IllegalStateException("BulkProcessingThresholdsHelper accessor already initialized!");
+                throw new IllegalStateException("BulkProcessingThresholds accessor already initialized!");
             }
 
             accessor = newAccessor;
@@ -371,7 +371,7 @@ public class ImplementationBridgeHelpers {
 
         public static BulkProcessingThresholdsAccessor getBulkProcessingThresholdsAccessor() {
             if (accessor == null) {
-                throw new IllegalStateException("BulkProcessingThresholdsHelper accessor is not initialized yet!");
+                throw new IllegalStateException("BulkProcessingThresholds accessor is not initialized yet!");
             }
 
             return accessor;