+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/HealthcareEntity.java
@@ -4,15 +4,17 @@
 package com.azure.ai.textanalytics.models;
 
 import com.azure.ai.textanalytics.implementation.HealthcareEntityPropertiesHelper;
+import com.azure.core.annotation.Immutable;
 import com.azure.core.util.IterableStream;
 
 /**
  * The {@link HealthcareEntity} model.
  */
+@Immutable
 public final class HealthcareEntity {
     private String text;
     private String normalizedText;
-    private String category;
+    private HealthcareEntityCategory category;
     private String subcategory;
     private double confidenceScore;
     private int offset;
@@ -33,7 +35,7 @@ public final class HealthcareEntity {
             }
 
             @Override
-            public void setCategory(HealthcareEntity healthcareEntity, String category) {
+            public void setCategory(HealthcareEntity healthcareEntity, HealthcareEntityCategory category) {
                 healthcareEntity.setCategory(category);
             }
 
@@ -94,7 +96,7 @@ public final class HealthcareEntity {
      *
      * @return The category value.
      */
-    public String getCategory() {
+    public HealthcareEntityCategory getCategory() {
         return this.category;
     }
 
@@ -161,7 +163,7 @@ public final class HealthcareEntity {
         this.normalizedText = normalizedText;
     }
 
-    private void setCategory(String category) {
+    private void setCategory(HealthcareEntityCategory category) {
         this.category = category;
     }