+++ sdk/spring/azure-spring-boot/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 3.7.0-beta.1 (Unreleased)
+
+
 ## 3.6.0 (2021-06-23)
 ### Breaking Changes
 - Deprecate aad.group.enable-full-list, use aad.group.allowed-group-ids=all instead.